Make modules_from_adjacencies more robust making sure the expression matrix consists of floating point numbers. This requirement might be violated when dealing with raw counts as input and is important when calculating modules with masking for dropouts.

+++ src/pyscenic/utils.py
@@ -243,7 +243,9 @@ def modules_from_adjacencies(adjacencies: pd.DataFrame,
 
     # Duplicate genes need to be removed from the expression matrix to avoid lookup problems in the correlation
     # matrix.
-    ex_mtx = ex_mtx.T[~ex_mtx.columns.duplicated(keep='first')].T
+    # In addition, also make sure the expression matrix consists of floating point numbers. This requirement might
+    # be violated when dealing with raw counts as input.
+    ex_mtx = ex_mtx.T[~ex_mtx.columns.duplicated(keep='first')].T.astype(float)
 
     # To make the pySCENIC code more robust to the selection of the network inference method in the first step of
     # the pipeline, it is better to use percentiles instead of absolute values for the weight thresholds.
